requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.7
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.7
daphne==4.2.1
dataclasses-json==0.6.7
distro==1.9.0
Django==6.0.4
djangorestframework==3.17.1
frozenlist==1.8.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.30
marshmallow==3.26.2
msgpack==1.1.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.4
openai==2.31.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
propcache==0.4.1
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
pyOpenSSL==26.0.0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
service-identity==24.2.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlparse==0.5.5
tavily-python==0.7.23
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
Twisted==25.5.0
txaio==25.12.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ujson==5.12.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
yarl==1.23.0
zope.interface==8.3
zstandard==0.25.0
